requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
databases
psycopg2-binary
python-dotenv
pydantic